fix: restructuring and removal of surplus
